Fixed build for later gcc versions

+++ zcm/transport/udpm/buffers.cpp
@@ -118,16 +118,16 @@ FragBuf *MessagePool::lookupFragBuf(struct sockaddr_in *key)
     return nullptr;
 }
 
-void MessagePool::_removeFragBuf(int index)
+void MessagePool::_removeFragBuf(unsigned int index)
 {
-    assert(0 <= index && index < (int)fragbufs.size());
+    assert(0 <= index && index < (unsigned int)fragbufs.size());
 
     // Update the total_size of the fragment buffers
     FragBuf *fbuf = fragbufs[index];
     totalSize -= fbuf->buf.size;
 
     // delete old element, move last element to this slot, and shrink by 1
-    size_t lastIdx = fragbufs.size()-1;
+    unsigned int lastIdx = (unsigned int)fragbufs.size()-1;
     fragbufs[index] = fragbufs[lastIdx];
     fragbufs.resize(lastIdx);
 
@@ -139,7 +139,7 @@ void MessagePool::removeFragBuf(FragBuf *fbuf)
 {
     // NOTE: this is kinda slow...
     // Search for the fragbuf index
-    for (int idx = 0; idx < (int)fragbufs.size(); idx++)
+    for (unsigned int idx = 0; idx < fragbufs.size(); idx++)
         if (fragbufs[idx] == fbuf)
             return this->_removeFragBuf(idx);
